feat(edit): cleanup TileCard
