+++ next-tavla/app/(admin)/edit/[id]/components/TileCard/PublicCode.tsx
@@ -1,7 +1,3 @@
 import classes from './styles.module.css'
 
-function PublicCode({ publicCode }: { publicCode: string | null }) {
-    return <div className={classes.publicCode}>{publicCode}</div>
-}
-
 export { PublicCode }
