+++ next-tavla/app/(admin)/edit/[id]/components/TileCard/styles.module.css
@@ -1,3 +1,12 @@
+.card {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background: var(--secondary-background-color);
+    border-radius: 0.5em;
+    padding: 1em;
+}
+
 .publicCode {
     background-color: var(--tertiary-background-color);
     padding: 0.35em 0.5em;
