--- next-tavla/app/(admin)/edit/[id]/components/styles.module.css
@@ -1,8 +1,0 @@
-.card {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    background: var(--secondary-background-color);
-    border-radius: 0.5em;
-    padding: 1em;
-}
